making sure values are floats

+++ src/ufdl/joblauncher/objdet/confidence.py
@@ -57,11 +57,11 @@ class ScoreDist(AbstractConfidenceScore):
         """
         result = dict()
         np_scores = np.array(scores)
-        result[self.name() + "-min"] = np.min(np_scores)
-        result[self.name() + "-max"] = np.max(np_scores)
-        result[self.name() + "-mean"] = np.mean(np_scores)
-        result[self.name() + "-median"] = np.median(np_scores)
-        result[self.name() + "-stdev"] = np.std(np_scores)
+        result[self.name() + "-min"] = float(np.min(np_scores))
+        result[self.name() + "-max"] = float(np.max(np_scores))
+        result[self.name() + "-mean"] = float(np.mean(np_scores))
+        result[self.name() + "-median"] = float(np.median(np_scores))
+        result[self.name() + "-stdev"] = float(np.std(np_scores))
         return result
 
 
@@ -91,11 +91,11 @@ class ObjectDims(AbstractConfidenceScore):
         :type values: list
         """
         np_values = np.array(values)
-        all[prefix + "-min"] = np.min(np_values)
-        all[prefix + "-max"] = np.max(np_values)
-        all[prefix + "-mean"] = np.mean(np_values)
-        all[prefix + "-median"] = np.median(np_values)
-        all[prefix + "-stdev"] = np.std(np_values)
+        all[prefix + "-min"] = float(np.min(np_values))
+        all[prefix + "-max"] = float(np.max(np_values))
+        all[prefix + "-mean"] = float(np.mean(np_values))
+        all[prefix + "-median"] = float(np.median(np_values))
+        all[prefix + "-stdev"] = float(np.std(np_values))
 
     def calculate(self, annotations, scores):
         """
